Refactors MarketerResource form schema and adds affiliate link generation
Updates the MarketerResource form schema in the Filament/Resources/MarketerResource.php file to include the autocapitalize attribute for the 'name' field and adds a new 'affiliate_link' field. Generates a unique referral code and affiliate link when the 'Generate Link' action is triggered. Updates the table view in the same file to display the 'affiliate_link' column. Also modifies the ApplicationController getCourses method in the Http/Controllers/ApplicationController.php file to pass the referral code from the URL to the view. Updates the Marketer model in the Models/Marketer.php file to remove the 'affiliate_link' field from the fillable array and use the guarded array instead. Modifies the create_marketers_table migration in the database/migrations/2025_01_18_064338_create_marketers_table.php file to add the 'affiliate_link' column to the 'marketers' table. Updates the application-form.blade.php view in the resources/views/enroll/ directory to pre-fill the 'referral_code' field with the value from the URL and make it read-only if pre-filled.